PowerDynamicsOpenAPIModels

The dynamics domain: machine, exciter, governor, and inverter-control models. A dynamic model attaches to a static injector through that injector's dynamic_injector id.

PowerDynamicsOpenAPIModels.ActiveRenewableControllerAB — Type

ActiveRenewableControllerAB

Parameters of Active Power Controller including REPCA1 and REECB1.

bus_control: ACBus identification number for voltage control. 0 identifies the local bus connected to this component.

from_branch_control: Monitored branch FROM bus number for line drop compensation (if 0 generator power will be used).

to_branch_control: Monitored branch TO bus number for line drop compensation (if 0 generator power will be used).

branch_id_control: Branch circuit id for line drop compensation. If 0 generator power will be used. in psy this is a string representing the branch name

freq_flag: Frequency Flag for REPCA1: false: disable, true: enable.

k_pg: Active power PI control proportional gain.

k_ig: Active power PI control integral gain.

t_p: Real power measurement filter time constant. Units: s.

fdbd_pnts: Frequency error dead band thresholds (fdbd1, fdbd2).

fe_lim: Upper/Lower limit on frequency error (fe_min, fe_max).

p_lim: Upper/Lower limit on power reference (P_min, P_max).

t_g: Power Controller lag time constant. Units: s.

d_dn: Droop for over-frequency conditions.

d_up: Droop for under-frequency conditions.

dp_lim: Upper/Lower limit on power reference ramp rates (dP_min, dP_max).

p_lim_inner: Upper/Lower limit on power reference for REECB (P_min_inner, P_max_inner).

t_pord: Power filter time constant REECB time constant. Units: s.

p_ref: Reference Power Set-point.

PowerDynamicsOpenAPIModels.RECurrentControlB — Type

RECurrentControlB

Parameters of the Inner Control part of the REECB model in PSS/E.

q_flag: Q Flag used for I_qinj.

pq_flag: PQ Flag used for the Current Limit Logic.

vdip_lim: Limits for Voltage Dip Logic (Vdip, Vup).

t_rv: Voltage Filter Time Constant. Units: s.

dbd_pnts: Voltage error deadband thresholds (dbd1, dbd2).

k_qv: Reactive current injection gain during over and undervoltage conditions.

iqinj_lim: Limits for Iqinj (I_qh1, I_ql1).

v_ref0: User defined reference. If 0, PowerSimulationsDynamics.jl initializes to initial terminal voltage.

k_vp: Voltage regulator proportional gain (used when QFlag = 1).

k_vi: Voltage regulator integral gain (used when QFlag = 1).

t_iq: Time constant for low-pass filter for state q_V when QFlag = 0. Units: s.

i_max: Maximum limit on total converter current.

PowerDynamicsOpenAPIModels.ReactiveRenewableControllerAB — Type

ReactiveRenewableControllerAB

Parameters of Reactive Power Controller including REPCA1 and REECB1.

bus_control: ACBus identification number for voltage control. 0 identifies the local bus connected to this component.

from_branch_control: Monitored branch FROM bus identification number for line drop compensation (if 0 generator power will be used).

to_branch_control: Monitored branch TO bus identification number for line drop compensation (if 0 generator power will be used).

branch_id_control: Branch circuit id for line drop compensation (as a string). If 0 generator power will be used.

vc_flag: Voltage Compensator Flag for REPCA1.

ref_flag: Flag for Reactive Power Control for REPCA1. false: Q-control, true: V-control.

pf_flag: Flag for Power Factor Control for Outer Control of REECB1. false: Q-control, true: Power Factor Control.

v_flag: Flag for Voltage Control for Outer Control of REECB1. false: Voltage Control, true: Q-Control.

t_fltr: Voltage or Q-power of REPCA Filter Time Constant. Units: s.

k_p: Reactive power PI control proportional gain.

k_i: Reactive power PI control integral gain.

t_ft: Reactive power lead time constant. Units: s.

t_fv: Reactive power lag time constant. Units: s.

v_frz: Voltage below which state xiq_oc (integrator state) is freeze.

r_c: Line drop compensation resistance (used when VC_Flag = 1).

x_c: Line drop compensation reactance (used when VC_Flag = 1).

k_c: Reactive current compensation gain (used when VC_Flag = 0).

e_lim: Upper/Lower limit on Voltage or Q-power deadband output (e_min, e_max).

dbd_pnts: Voltage or Q-power error dead band thresholds (dbd1, dbd2).

q_lim: Upper/Lower limit on reactive power V/Q control in REPCA (Q_min, Q_max).

t_p: Active power lag time constant in REECB. Used only when PF_Flag = 1. Units: s.

q_lim_inner: Upper/Lower limit on reactive power input in REECB (Q_min_inner, Q_max_inner). Only used when V_Flag = 1.

v_lim: Upper/Lower limit on reactive power PI controller in REECB (V_min, V_max). Only used when V_Flag = 1.

k_qp: Reactive power regulator proportional gain (used when V_Flag = 1).

k_qi: Reactive power regulator integral gain (used when V_Flag = 1).

q_ref: Reference Reactive Power Set-point.

v_ref: Reference Voltage Set-point.

PowerDynamicsOpenAPIModels.RenewableEnergyConverterTypeA — Type

RenewableEnergyConverterTypeA

Parameters of a renewable energy generator/converter model, this model corresponds to REGCA1 in PSSE.

t_g: Converter time constant. Units: s.

rrpwr: Low Voltage Power Logic (LVPL) ramp rate limit.

brkpt: LVPL characteristic voltage 2.

zerox: LVPL characteristic voltage 1.

lvpl1: LVPL gain.

vo_lim: Voltage limit for high voltage reactive current management.

lv_pnts: Voltage points for low voltage active current management (Lvpnt0, Lvpnt1).

io_lim: Current limit for high voltage reactive current management (specified as a negative value).

t_fltr: Voltage filter time constant for low voltage active current management. Units: s.

k_hv: Overvoltage compensation gain used in the high voltage reactive current management.

iqr_lims: Limit on rate of change for reactive current (Iqrmin, Iqrmax).

accel: Acceleration factor.

lvpl_sw: Low voltage power logic (LVPL) switch. (false: LVPL not present, true: LVPL present).

q_ref: Initial condition of reactive power from power flow.

r_source: Output resistor used for the Thevenin Equivalent.

x_source: Output reactance used for the Thevenin Equivalent.

PowerDynamicsOpenAPIModels.RoundRotorMachine — Type

RoundRotorMachine

Parameters of 4-states round-rotor synchronous machine with quadratic/exponential saturation: IEEE Std 1110 5.3.2 (Model 2.2). GENROU or GENROE model in PSSE and PSLF.

id: Unique integer identifier for this component.

r: Armature resistance.

td0_p: Time constant of transient d-axis voltage. Units: s.

td0_pp: Time constant of sub-transient d-axis voltage. Units: s.

tq0_p: Time constant of transient q-axis voltage. Units: s.

tq0_pp: Time constant of sub-transient q-axis voltage. Units: s.

xd: Reactance after EMF in d-axis.

xq: Reactance after EMF in q-axis.

xd_p: Transient reactance after EMF in d-axis.

xq_p: Transient reactance after EMF in q-axis.

xd_pp: Sub-Transient reactance after EMF in d-axis. Note: Xdpp = Xqpp.

xl: Stator leakage reactance.

se: Saturation factor at 1 and 1.2 pu flux: S(1.0) = B(|psi_pp|-A)^2.

gamma_d1: Do not modify

gamma_q1: Do not modify

gamma_d2: Do not modify

gamma_q2: Do not modify

gamma_qd: Do not modify

PowerDynamicsOpenAPIModels.SEXS — Type

SEXS

Parameters of Simplified Excitation System Model - SEXS in PSSE.

id: Unique integer identifier for this component.

ta_tb: Ratio of lead and lag time constants.

tb: Lag time constant. Units: s.

k: Gain.

te: Field circuit time constant. Units: s.

v_lim: Field voltage limits.

v_ref: Reference Voltage Set-point.

PowerDynamicsOpenAPIModels.SteamTurbineGov1 — Type

SteamTurbineGov1

Steam Turbine-Governor. This model considers both TGOV1 or TGOV1DU in PSS/E.

id: Unique integer identifier for this component.

r: Droop parameter.

t1: Governor time constant. Units: s.

valve_position_limits: Valve position limits.

t2: Lead Lag Lead Time constant. Units: s.

t3: Lead Lag Lag Time constant. Units: s.

d_t: Turbine Damping.

db_h: Deadband for overspeed.

db_l: Deadband for underspeed.

t_rate: Turbine Rate. If zero, generator base is used. Units: MW.

p_ref: Reference Power Set-point.